Moved list stuff over to the visitor trait
